Uploaded optimized PNG of TM3D facility render.
I was testing oxypng utility and I needed some PNGs that I could use to see what it can do. As the result, this particular PNG was a neat test subject.

I managed to reduce the file size of the image and add interlacing so it could display instantly and then gain more detail during the load process.